remove unnecessary type conversions

+++ template/funcs.go
@@ -972,7 +972,7 @@ func loop(ifaces ...interface{}) (<-chan int64, error) {
 		switch v.Kind() {
 		case reflect.Int, reflect.Int8, reflect.Int16, reflect.Int32,
 			reflect.Int64:
-			return int64(v.Int()), nil
+			return v.Int(), nil
 		case reflect.Uint, reflect.Uint8, reflect.Uint16, reflect.Uint32,
 			reflect.Uint64:
 			return int64(v.Uint()), nil
